fix: remove duplicate code to create native envs
